add balance

+++ Stock.Analysis.0607/Models/TestCase.cs
@@ -2,7 +2,7 @@
 {
     public class TestCase
     {
-        public decimal Funds { get; set; }
+        public double Funds { get; set; }
         public int ShortTermMa { get; set; }
         public int LongTermMa { get; set; }
     }
